+++ node/DLNode/node_face_detection.py
@@ -113,7 +113,7 @@ class Node(Node):
         use_gpu = self._opencv_setting_dict['use_gpu']
 
 
-        black_image = np.zeros((small_window_w, small_window_h, 3))
+        black_image = np.zeros((small_window_w, small_window_h, 3), dtype=np.uint8)
         black_texture = self.convert_cv_to_dpg(
             black_image,
             small_window_w,
